pyproject.toml
[tool.poetry]
name = "drf-triad-permissions"
version = "0.3.0"
description = "Django Rest Framework viewset permissions through triads"
license = "MIT"
authors = ["Lorenzo Peña <lorinkoz@gmail.com>"]
readme = "README.rst"
repository = "https://github.com/lorinkoz/drf-triad-permissions"
keywords = ["django", "djangorestframework", "permissions"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 2.0",
    "Framework :: Django :: 2.1",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
]

[tool.poetry.dependencies]
python = "^3.6"
django = ">=2.0,<4.0"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = {extras = ["toml"], version = "^5.1"}
flake8 = "^3.8.4"
flake8-bugbear = "^20.11.1"
flake8-comprehensions = "^3.3.1"
flake8-no-types = "^1.1.1"
flake8-tidy-imports = "^4.2.1"
isort = "^5.6.4"

[tool.black]
line-length = 120

[tool.isort]
atomic = true
profile = "black"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_django = ["django", "rest_framework"]
combine_as_imports = true
line_length = 120

[tool.coverage.run]
source = ["drf_triad_permissions"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
